requirements.txt
Django==1.10.4
django-bootstrap3==7.0.1
django-haystack==2.6
whitenoise==3.2.2
dj-database-url==0.4.1
gunicorn==19.6.0
psycopg2==2.6.1
Whoosh==2.7.4
